% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nu_aero.R
\name{calc_nu_aero}
\alias{calc_nu_aero}
\title{Calculate probability of infection via aerosol \eqn{\nu^{AERO}}}
\usage{
calc_nu_aero(
  AER = NULL,
  s = NULL,
  lambda = NULL,
  C_nu = NULL,
  C_i = NULL,
  IR = NULL,
  ER = NULL,
  V_d = NULL,
  V_air = NULL,
  t_contact = NULL,
  r = NULL,
  nsamples = NULL,
  print_pars = FALSE,
  seed = NULL
)
}
\arguments{
\item{AER}{air exchange in \eqn{hr^{-1}}}

\item{s}{settling rate; \eqn{hr^{-1}}}

\item{lambda}{innactivation rates \eqn{hr^{-1}}}

\item{C_nu}{viral load in sputum; RNA copies/ml}

\item{C_i}{conversion factor for quanta/RNA copy}

\item{IR}{inhalation rate; \eqn{m^3/hr}}

\item{ER}{exhalation rate; \eqn{m^3/hr}}

\item{V_d}{exhaled droplet volume concentration; ml exhaled droplets/ \eqn{m^3}}

\item{V_air}{fixed volume; \eqn{m^3}}

\item{t_contact}{time of contact with contaminated airspace (hr)}

\item{r}{species-specific probability of infection from 1 quantum. Default is for r_deer with expert elicited values}

\item{nsamples}{default to 1, but if specified > 1 will draw \code{nsamples} from the default distributions of parameters}

\item{print_pars}{logical. If TRUE, will return a list with all parameters used to make calculations}

\item{seed}{if setting a seed, specify number}
}
\value{
returns a numeric vector of length equal to nsamples or length of vectors provided in arguments
}
\description{
\code{calc_nu_aero} will calculate one value for probability of infection using defaults if no arguments are provided. Defaults are described in the sir_model_description document and are sourced from the literature or expert elicitation.
}
\examples{
calc_nu_aero() # will return a single value using defaults
}
