% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nu_dc.R
\name{calc_nu_dc}
\alias{calc_nu_dc}
\title{Calculate transmission for direct contact}
\usage{
calc_nu_dc(
  V_DC = NULL,
  C_nu = NULL,
  pfu_conv = NULL,
  k = NULL,
  nsamples = NULL,
  seed = NULL
)
}
\arguments{
\item{seed}{}
}
\description{
Calculate transmission for direct contact
}
