% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sigma_dc.R
\name{calc_sigma_dc}
\alias{calc_sigma_dc}
\title{Calculate probability of infection given physical contact
\eqn{\sigma^{DC}}}
\usage{
calc_sigma_dc(
  V_DC = NULL,
  C_nu = NULL,
  pfu_conv = NULL,
  k = NULL,
  nsamples = NULL,
  seed = NULL
)
}
\arguments{
\item{V_DC}{Transferred volume (ml). Default is 0.1}

\item{C_nu}{viral load in sputum; RNA copies/ml. Default samples from expert
elicited distribution of parameter 'Viral Load'. Concentration of
infectious particles in sputum (gc per ml)}

\item{pfu_conv}{conversion value (PFU per GC). Set to 1/10^5.2}

\item{k}{set to 410 as the default}

\item{nsamples}{number of samples wanted. Default is set to 1}

\item{seed}{sets a seed for random draws of parameter from default
distributions.}
}
\value{
returns a number or vector of probability values for probability of
infection via direct contact.
}
\description{
\code{calc_sigma_dc} will calculate one value for probability of
infection using
defaults if no arguments are provided. Defaults are described in the
sir_model_description document and are sourced from the literature or expert
elicitation.
}
\details{
Mathematical background to calculate this infection probability
given physical contact. We use a Wells-Riley dose response model that
estimates infection probability as a log-logistic function of dose
(\eqn{d^{dc}}; plaque-forming units, PFU) and the dose-response function
(\eqn{k}; Watanabe et al. 2010). The dose received is the product of a
typical volume of sputum transferred given contact (\eqn{V^{DC}}) and the
concentration of SARS-CoV-2 in sputum (\eqn{C_{\nu}}). We initially assume
that at each contact \eqn{100\mu l} of sputum is transferred between
individuals making contact.
\deqn{\sigma^{DC} = 1 - e^{-(d^{DC}/k)}} where
\deqn{d^{DC} = V^{DC} \cdot C_{\nu} \cdot pfuConv}
}
\examples{
\dontrun{
calc_sigma_dc() # will run the defaults
calc_sigma_dc(nsamples = 10) # use defaults with a random draw for values of
C_nu from expert elicitation}

}
