% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elicitation_data}
\alias{elicitation_data}
\title{elicitation_data}
\format{
\subsection{\code{elicitation_data}}{

A data frame with 4 columns and 15 rows, each row corresponding to an
elicited parameter:
\describe{
\item{parameter}{Parameter elicited}
\item{mu}{mean value}
\item{sd}{standard deviation}
\item{family}{distribution family for the parameters given}
}
}
}
\source{
Contact: Elias Rosenblatt erosenbl@uvm.edu
}
\usage{
elicitation_data
}
\description{
Group averaged (using quantile averaging) parameter estimates
for parameters derived using expert elicitation.
}
\details{
Parameters include 1) Days until a recovered individual can be
reinfected with SARS-CoV-2 (Parameter = Temporary Immunity; units = days), 2)
Number of individuals infected via wastewater for every 100 individuals
infected via direct transmission (Parameter = Wastewater Infections; units =
individuals), 3) Number of individuals infected via formites for every 100
individuals infected via direct transmission (Parameter = Fomite Infections;
units = individuals), 4) Ratio of deer sputum viral load : humansputum viral
load (Parameter = Viral Load; units = multiplicative ratio), 5) Dose-response
value for quantum dosage using the Wells-Riley dose response model (Parameter
= Dose-Response; units = r value). 6) deer-deer
contact duration (minutes), 7) deer-deer direct contact probability, 8)
deer-deer proximity rate with baiting (per day; 17 events per day without
baiting), 9) deer-human proximity rate, rural (per 120 days), 10) deer-human
proximity duration, rural (minutes), 11) deer-human proximity rate, suburban
(per 120 days), 12) deer-human proximity duration, suburban (minutes), 13)
deer-human proximity rate, captive (per 120 days), 14) deer-human proximity
duration, captive (minutes), 15) deer-deer proximity rate in captivity (per
day).
}
\examples{
\dontrun{
head(elicitation_data)
}

}
\keyword{datasets}
