% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run ODE Solver for SIRS model}
\usage{
run(
  iter = NULL,
  initial_compartments = example_inits,
  parameters = NULL,
  times = NULL,
  name = NULL
)
}
\arguments{
\item{iter}{Number of iterations, often referred to as 'nsamples' in package examples}

\item{initial_compartments}{Initial compartment sizes (proportions, stored as list)}

\item{parameters}{SIRS parameters, stored as a list}

\item{times}{Length of projection (days)}

\item{name}{Name of context being simulated}
}
\value{
#List containing iteration details, including initial compartment sizes, parameters, context name, and daily, proportional sizes of each SIRS compartment.
}
\description{
Run ODE Solver for SIRS model
}
\examples{
nsamples = 10

example_inits <- c(S_wild = 1, I_wild = 0,
                   R_wild = 0, S_captive = 1,
                   I_captive = 0, R_captive = 0)

# set the time to run
example_times <-  seq(0, 365, by = 1)
# Set parameters of transmission, immunity, recovery

example_params <- c(alpha_immunity = 0.03,
                    beta_aero_ww = 0.01,
                    beta_aero_cw = 0.01,
                    beta_aero_cc = 0.02,
                    beta_aero_hw = 0.01,
                    beta_aero_hc = 0.2,
                    beta_dc_ww = 0.01,
                    beta_dc_cw = 0.01,
                    beta_dc_cc = 0.01,
                    gamma_recov = 0.01,
                    I_human = 0.05,
                    boost = 0)

run(iter = nsamples, initial_compartments = example_inits, parameters = example_params, times = example_times, name = Test)
}
