% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sirs_projection.R
\name{run_sirs_projection}
\alias{run_sirs_projection}
\title{In progress Run SIRS projection}
\usage{
run_sirs_projection(
  nsamples = 100,
  seed = 123,
  wild_scenario = "rural",
  ndays = 120,
  type_contact = "low",
  ...
)
}
\arguments{
\item{nsamples}{number of samples to return}

\item{seed}{if specified will set a seed}

\item{wild_scenario}{deafult to 'rural' but can also calculate 'suburban'}

\item{ndays}{number of days for projection specified as 120}

\item{...}{}
}
\value{
returns a list column tibble with parameters and sirs projection.
}
\description{
This function is still in development but it is a wrapper for other functions in the package and will generate a projection for SIRS model using defaults for a specified number of samples
}
\examples{
run_sirs_projection(nsamples = 1, seed = 63)
}
