% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_sirs.R
\name{simple_sirs}
\alias{simple_sirs}
\title{Simple function to generate SIRS model projections}
\usage{
simple_sirs(time, state, parameters)
}
\arguments{
\item{time}{vector of values for which to generate the projection (length in days)}

\item{state}{initial state for compartment populations}

\item{parameters}{list of parameters to generate projection}
}
\value{
list of values at each time step
}
\description{
Simple function to generate SIRS model projections
}
