% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_sirs_with_cumulative.R
\name{simple_sirs_with_cumulative}
\alias{simple_sirs_with_cumulative}
\title{Simple function to generate SIRS model projections}
\usage{
simple_sirs_with_cumulative(time, state, parameters)
}
\arguments{
\item{time}{vector of values for which to generate the projection (length in
days).}

\item{state}{initial state for compartment populations. This should be a
named vector for starting values of S_wild, I_wild, R_wild, S_captive,
I_captive, R_captive.}

\item{parameters}{list of parameters to generate projection. The parameters
should include transmission parameters, immunity and recovery rates, and
proportion of infected humans.}
}
\value{
when used with the \code{deSolve::ode()} function, it will return a
dataframe with the proportion of individuals in each of the SIR
compartments at each time point, as well as a cumulative case count (proportion)
}
\description{
\code{simple_sirs_with_cumulative} is written to work alongside the \code{deSolve::ode()} function, and
it will return an object with the proportion of individuals found in each of
the SIR compartments at each of the specified time points.
}
\details{
\code{simple_sirs_with_cumulative()} cannot be run with \code{run_steady()} to estimate probability of persistence. Instead, use \code{simple_sirs()}, which does not return cumulative proportion of population infected.
}
\examples{
# prepare the input parameters:
example_inits <- c(S_wild = 1, I_wild = 0,
                   R_wild = 0, I_wild_cumulative = 0,S_captive = 1,
                   I_captive = 0, R_captive = 0, I_captive_cumulative = 0)

# set the time to run
example_times <-  seq(0, 500, by = 1)
# Set parameters of transmission, immunity, recovery

example_params <- c(alpha_immunity = 0.03,
                    beta_aero_ww = 0.01,
                    beta_aero_cw = 0.01,
                    beta_aero_cc = 0.02,
                    beta_aero_hw = 0.01,
                    beta_aero_hc = 0.2,
                    beta_dc_ww = 0.01,
                    beta_dc_cw = 0.01,
                    beta_dc_cc = 0.01,
                    gamma_recov = 0.01,
                    I_human = 0.05,
                    boost = 0)

# run the ode function:

deSolve::ode(y = example_inits, times = example_times, parms = example_params, func = whitetailedSIRS::simple_sirs_with_cumulative)


}
