\name{whitetailedSIRS}
\alias{whitetailedSIRS}

\title{whitetailedSIRS: A package to project SARS-CoV-2 outbreak dynamics in
white-tailed deer
}

\description{
This software release contains several R scripts that generate epidemic projections of SARS-CoV-2 in white tailed deer populations using a **S**usceptible-**I**nfected-**R**ecovered-**S**usceptible (SIRS) modeling framework. We provide a workflow of vignettes used in Rosenblatt et al. In Prep and Cook et al. In Prep. Users are able to specify transmission parameters for human-deer and deer-deer transmission to quantify changes in outbreak dynamics. The outputs of this package are ggplot friendly, and for the most part, we use a list-column workflow when working with multiple simulations. Details on the model can be found in the vignettes.
}


