% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternatives.R
\name{alternative}
\alias{alternative}
\title{alternative: a function to define parameters for solving white-tailed deer SIRS ODE equations}
\usage{
alternative(
  alpha_immunity = NULL,
  c_ww = NULL,
  c_cw = NULL,
  c_cc = NULL,
  c_hw = NULL,
  c_hc = NULL,
  nu_aero_deer_deer_wild = NULL,
  nu_aero_deer_deer_captive = NULL,
  nu_aero_deer_human_wild = NULL,
  nu_aero_deer_human_capt = NULL,
  sigma_dc = NULL,
  nu_dc_deer_deer = NULL,
  gamma_recov = NULL,
  I_human = NULL,
  boost = NULL
)
}
\arguments{
\item{alpha_immunity}{Inverse of duration of temporary immunity after entering recovered compartment (per day, 0-1)}

\item{c_ww}{Proximity rate (<=1.5m) between wild deer (proximity events per day)}

\item{c_cw}{Proximity rate (<=1.5m) between captive and wild deer along fencelines(proximity events per day)}

\item{c_cc}{Proximity rate (<=1.5m) between captive deer (proximity events per day)}

\item{c_hw}{Proximity rate (<=1.5m) between humans and wild deer (proximity events per day)}

\item{c_hc}{Proximity rate (<=1.5m) between humans and captive deer (proximity events per day)}

\item{nu_aero_deer_deer_wild}{Probability of infection via aerosol transmission in wild deer}

\item{nu_aero_deer_deer_captive}{Probability of infection via aerosol transmission in captive deer}

\item{nu_aero_deer_human_wild}{Probability of infection via aerosol transmission from humans to wild deer}

\item{nu_aero_deer_human_capt}{Probability of infection via aerosol transmission from humans to captive deer}

\item{sigma_dc}{Probability of direct contact between deer, given proximity}

\item{nu_dc_deer_deer}{Probability of infection via fluid transmission between deer}

\item{gamma_recov}{Duration of recovery from infection (days)}

\item{I_human}{Prevalence in human population (Proportion)}

\item{boost}{Boosting rate in captive deer (Proportion)}
}
\value{
Returns a list of vectors for each parameter
}
\description{
alternative: a function to define parameters for solving white-tailed deer SIRS ODE equations
}
\details{
alternative() creates a list of parameter values that will be fed into a SIRS ODE solver. The length of each item in the output is determined by the length of all arguments, which must be equal. If argument(s) is not filled with outputs from another function in this package (e.g. calc_contact_rate, calc_nu_aero, calc_nu_dc, draw_elicitation_samples), the user must fill argument with vector of values of the same length as the other arguments.
}
\examples{
alternative(alpha_immunity = c((1/30),(1/60),(1/90),(1/120),(1/150)), c_ww = rep(10,5), c_cw = rep(1,5), c_cc = rep(20,5), c_hw = rep(0.01,5), c_hc = rep(0.25,5), nu_aero_deer_deer_wild = rep(0.01,5), nu_aero_deer_deer_captive = rep(0.1,5), nu_aero_deer_human_wild = rep(0.001,5), nu_aero_deer_human_capt = rep(0.05,5), sigma_dc = rep(0.2,5), nu_dc_deer_deer = rep(0.02,5), gamma_recov = rep(1/6,5), I_human = rep(0.05,5), boost = rep(0,5))
}
