% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contact_rate.R
\name{calc_contact_rate}
\alias{calc_contact_rate}
\title{Contact rate calculation for wild deer}
\usage{
calc_contact_rate(
  kappa = NULL,
  N_w = NULL,
  q = NULL,
  A_w = NULL,
  rho_attractant = NULL,
  nsamples = NULL,
  seed = NULL,
  type_contact = NULL
)
}
\arguments{
\item{kappa}{value for contact scaling constant (influences slope of density-
contact relationship; reported in Habib et al. 2011)}

\item{N_w}{total population size of deer. Defaults is a random draw from a
Poisson distribution with mean = 1000.}

\item{q}{value for concavity scaling constant (0-1, with 0 equating to
density dependence and 1 equating to frequency dependence; reported in Habib
et al. 2011)}

\item{A_w}{area inhabited by N, in square kilometers. Default set to 100
square kilometers to match Habib et al.'s (2011) approach.}

\item{rho_attractant}{proportional change for attractants present in
simulated conditions (optional). Default set to 1}

\item{nsamples}{number of values to be returned. Default is 1, but will
return a vector if > 2}

\item{seed}{if specified, sets a seed for the function}

\item{type_contact}{option between "low", "medium", and "high" based on
parameters from Habib et al. 2011 for particular values of wooded habitat
available to the deer population in question. "low" contact fixes kappa to
16.37 and q to 0.53, mimicking proximity rates in areas with 53\% wooded
habitat. "med" contact fixes kappa to 11.35 and q to 0.34, mimicking
proximity rates in areas with 26\% wooded habitat. "high" contact fixes kappa
to 15.58 and q to 0.32, mimicking proximity rates in areas with 12\% wooded
habitat. If set to "manual", manual input of kappa and q are needed. All
values from Habib et al. (2011) are available in the \code{contact_rate_params}
dataset.}
}
\value{
returns a number or numeric vector if nsamples > 1
}
\description{
\code{calc_contact_rate()} uses proximity rate model developed by
Habib et al. (2011) to estimate proximity rates for deer in wild settings,
conditional on density and habitat availability. This function can also be
used for captive settings with conditions that result in identical deer-deer
proximity rates. The user can alter proximity rates with the inclusion of
attractants on a landscape.
}
\examples{
 \dontrun{
calc_contact_rate(type_contact = "low")
calc_contact_rate(rho_attractant = 1, kappa = 16.37, N_w = 1000, q = 0.53, A_w = 100, type_contact = "manual")
calc_contact_rate(c(1,1), c(16.37, 16.37), rpois(2, 1000), c(0.53, 0.53), c(100, 100), type_contact = "manual")}

}
