% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_contact_rate.R
\name{calc_contact_rate}
\alias{calc_contact_rate}
\title{Contact rate calculation for wild deer}
\usage{
calc_contact_rate(
  sigma_season = NULL,
  scaling_c = NULL,
  N_w = NULL,
  q = NULL,
  A_w = NULL,
  nsamples = NULL,
  seed = NULL,
  type_contact = NULL
)
}
\arguments{
\item{sigma_season}{value for season scaling (derived from Williams et al 2014; Appendix A). Default set to 1}

\item{scaling_c}{value for contact scaling constant (influences slope of density-contact relationship; reported in Habib et al. 2011; Appendix A)}

\item{N_w}{total population size of deer. Poisson dist w/ean = 10}

\item{q}{value for concavity scaling constant (0-1, with 0 equating to density dependence and 1 equating to frequency dependence; reported in Habib et al. 2011, Appendix A)}

\item{A_w}{area inhabited by N}

\item{nsamples}{number of values to be returned. Default is 1, but will return a vector if > 2}

\item{seed}{if specified, sets a seed for the function}

\item{type_contact}{option between "low", "medium", and "high" based on parameters from Habib 2011 table. If set to NULL, manual input of scaling_c and q are needed.}
}
\value{
returns a number or numeric vector if nsamples > 1
}
\description{
Contact rate calculation for wild deer
}
\examples{
calc_contact_rate()
calc_contact_rate(1, 16.37, 1000, 0.53, 100)
calc_contact_rate(c(1,1), c(16.37, 16.37), rpois(2, 1000), c(0.53, 0.53), c(100, 100))
}
