% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_nu_aero.R
\name{calc_nu_aero}
\alias{calc_nu_aero}
\title{Calculate probability of infection via aerosol \eqn{\nu^{AERO}}}
\usage{
calc_nu_aero(
  AER = NULL,
  s = NULL,
  lambda = NULL,
  C_nu = NULL,
  C_i = NULL,
  IR = NULL,
  ER = NULL,
  V_d = NULL,
  V_air = NULL,
  t_contact = NULL,
  r = NULL,
  nsamples = NULL,
  seed = NULL
)
}
\arguments{
\item{AER}{air exchange rate \eqn{hr^{-1}}. Default is set to \eqn{4 hr^{-1}}}

\item{s}{settling rate; \eqn{hr^{-1}} Default is set to \eqn{0.24 hr^{-1}}}

\item{lambda}{inactivation rates \eqn{hr^{-1}}. Default is set to \eqn{0.63 hr^{-1}}}

\item{C_nu}{viral load in sputum; RNA copies/ml. Default samples from expert
elicited distribution of parameter 'Viral Load'}

\item{C_i}{conversion factor for quanta/RNA copy. Default set to 0.0014}

\item{IR}{inhalation rate; \eqn{m^3/hr}.Default is set to 0.846 for deer}

\item{ER}{exhalation rate; \eqn{m^3/hr}. Default IR = ER}

\item{V_d}{exhaled droplet volume concentration; ml exhaled droplets/
\eqn{m^3}. Default to 0.009}

\item{V_air}{fixed volume; \eqn{m^3}. Default to \eqn{7.07m^3}, corresponding to a half-sphere with a 1.5m radius}

\item{t_contact}{time of contact with contaminated airspace (hr).}

\item{r}{species-specific probability of infection from 1 quantum. Default is
for r_deer with expert elicited values.}

\item{nsamples}{default to 1, but if specified > 1 will draw \code{nsamples} from
the default distributions of parameters}

\item{seed}{if setting a seed, specify number}
}
\value{
a numeric vector of length equal to nsamples for values of nu_aero
}
\description{
\code{calc_nu_aero} will calculate one value for probability of infection using
defaults if no arguments are provided. Defaults are described in the
sir_model_description document and are sourced from the literature or expert
elicitation.
}
\details{
Mathematical background for this calculation. An infected
individual emits viral particles at a particular rate \eqn{ER_q} in
quanta/hr as the
product of the arguments described above: \deqn{ER_q = C_{\nu} \cdot C_i
  \cdot ER \cdot V_d} This is used to model the instantaneous concentration
of viral particles (C) in well-mixed air space (quanta/\eqn{m^3}) around an
infected individual as follows: \deqn{C = \frac{ER_q}{IVRR \cdot V_{air}}}
where the loss rate (IVRR) is given by: \deqn{AER + s + \lambda} When a
susceptible individual enters the contaminated airspace surrounding an
infected individual, the dose (\eqn{Q_A}) is the product of inhalation
rate, concentration of viral particles, and time of contact: \deqn{Q_A = IR
  \cdot C \cdot t_{contact}} The dose \eqn{Q_A} is converted into a
probability of infection using the Wells-Riley infection model as a
function of the dose received and a species-specific probability of
infection from 1 quantum. \deqn{\nu^{AERO} = 1 - e^{-rQ}}
}
\examples{
calc_nu_aero() # will return a single value using defaults
calc_nu_aero(nsamples = 10)
}
