% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_elicitation_samples.R
\name{draw_elicitation_samples}
\alias{draw_elicitation_samples}
\title{Random draw of parameters from expert elicitation}
\usage{
draw_elicitation_samples(
  elicitation_data,
  nsamples,
  seed = NULL,
  return_df = FALSE
)
}
\arguments{
\item{elicitation_data}{defaults to the package data on expert elicitation}

\item{nsamples}{a number of samples to draw}

\item{seed}{if specified, a seed is set}

\item{return_df}{is TRUE the object returned is a data frame}
}
\value{
returns a tibble with list column for the samples drawn unless
otherwise specified in the \code{return_df} argument.
}
\description{
Random draw of parameters from expert elicitation
}
