% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_projections}
\alias{scenario_projections}
\title{scenario_projections}
\format{
\code{scenario_projections}
}
\source{
Derived using \code{SIRS_analysis_by_contexts.Rmd}
}
\usage{
scenario_projections
}
\description{
A tibble containing projected SIR compartment sizes for every simulation run. Columns in this tibble include the scenario (Context), scenario run (run_id), and a list containing daily (time) compartment sizes for wild and captive SIR compartments (S_wild, I_wild, R_wild, & S_captive, I_captive, R_captive) and cumulative infections (I_wild_cumulative and I_captive_cumulative).
}
\keyword{datasets}
